Split out programme helpers
