+++ .prettierignore
@@ -1,4 +1,5 @@
-*.json
+package.json
+package-lock.json
 node_modules/
 tmp/
 public/
